backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
ultralytics
sentence-transformers
chromadb
Pillow
numpy
torch
openai